# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: De Lá do Pará API
  version: 0.0.1
  description: >
    Contrato canônico da API (C11). Toda operação HTTP implementada precisa
    estar aqui antes do código; o cliente Angular é gerado deste arquivo e
    nunca editado manualmente. Endpoints de negócio chegam com as specs
    (C15+); este seed prova o gate.
servers:
  - url: http://localhost:8080
    description: API local
security:
  - sessionCookie: []
paths:
...
  /api/v1/checkout/snapshots:
    post:
      operationId: startCheckoutSnapshot
      summary: Congelar o carrinho convidado para o checkout
      security: []
      responses:
        '200':
          description: Identidade do snapshot criado
          content:
            application/json: {schema: {$ref: '#/components/schemas/CheckoutSnapshot'}}
        '404': {$ref: '#/components/responses/Problem404'}
        '422': {$ref: '#/components/responses/Problem422'}
...
  /api/v1/checkout/{snapshotId}/pickup-options:
    get:
      operationId: listCheckoutPickupOptions
      summary: Consultar opções de retirada do snapshot
      security: []
      parameters:
        - {in: path, name: snapshotId, required: true, schema: {type: string, format: uuid}}
        - {in: query, name: snapshotVersion, required: true, schema: {type: integer, format: int64, minimum: 0}}
      responses:
        '200':
          description: Estado e opções de retirada server-owned
          content:
            application/json: {schema: {$ref: '#/components/schemas/PickupOptions'}}
        '404': {$ref: '#/components/responses/Problem404'}
        '409': {$ref: '#/components/responses/Problem409'}
  /api/v1/checkout/{snapshotId}/pickup-selection:
    post:
      operationId: selectCheckoutPickupOption
      summary: Validar e selecionar uma opção de retirada
      security: []
      parameters:
        - {in: path, name: snapshotId, required: true, schema: {type: string, format: uuid}}
        - {in: query, name: snapshotVersion, required: true, schema: {type: integer, format: int64, minimum: 0}}
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/PickupSelectionRequest'}}
      responses:
        '200':
          description: Opção de retirada validada para o snapshot
          content:
            application/json: {schema: {$ref: '#/components/schemas/PickupSelection'}}
        '404': {$ref: '#/components/responses/Problem404'}
        '409': {$ref: '#/components/responses/Problem409'}
...
components:
  responses:
...
    Problem409:
      description: Conflito de versão
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemDetail'}}}
    Problem404:
      description: Recurso não encontrado
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemDetail'}}}
    Problem422:
      description: Operação sem modalidade compatível
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemDetail'}}}
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DLSESSION
      description: >
        Sessão Spring Security persistida no PostgreSQL (D64), cookie
        HttpOnly, Secure e SameSite=Lax.
  schemas:
...
    ProblemDetail:
      description: Erro padronizado (RFC 9457 + código comercial estável)
      type: object
      required: [title, status, codigo]
      example:
        title: Erro
        status: 409
        detail: identificador já cadastrado
        codigo: CATALOG_002
        correlationId: 4b1e7c69-628c-4c81-92f7-f30d6821d30a
      properties:
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        codigo:
          type: string
          description: Código comercial estável (ex. CHECKOUT_CONFLITO)
          example: RECURSO_NAO_ENCONTRADO
        correlationId:
          type: string
          format: uuid
...
    CheckoutSnapshot:
      type: object
      additionalProperties: false
      required: [snapshotId, snapshotVersion]
      properties:
        snapshotId: {type: string, format: uuid}
        snapshotVersion: {type: integer, format: int64, minimum: 0}
...
    PickupOptions:
      type: object
      additionalProperties: false
      required: [status, options, unavailableSkuIds]
      properties:
        status: {type: string, enum: [AVAILABLE, UNAVAILABLE]}
        options:
          type: array
          items: {$ref: '#/components/schemas/PickupOption'}
        unavailableSkuIds:
          type: array
          items: {type: string, format: uuid}
    PickupOption:
      type: object
      additionalProperties: false
      required: [id, point, window, preparationDays]
      properties:
        id: {type: string}
        point: {type: string}
        window: {type: string}
        preparationDays: {type: integer, minimum: 0}
...
    PickupSelectionRequest:
      type: object
      additionalProperties: false
      required: [pickupOptionId]
      properties:
        pickupOptionId: {type: string, minLength: 1}
    PickupSelection:
      type: object
      additionalProperties: false
      required: [id, point, window, preparationDays]
      properties:
        id: {type: string}
        point: {type: string}
        window: {type: string}
        preparationDays: {type: integer, minimum: 0}
...